rename LTK to Pair
