--- omnipod-dash/src/main/java/info/nightscout/androidaps/plugins/pump/omnipod/dash/driver/comm/ltk/LTK.kt
@@ -1,7 +1,0 @@
-package info.nightscout.androidaps.plugins.pump.omnipod.dash.driver.comm.ltk
-
-data class LTK(val ltk: ByteArray) {
-    init {
-        require(ltk.size == 16)
-    }
-}
