+++ omnipod-dash/src/main/java/info/nightscout/androidaps/plugins/pump/omnipod/dash/driver/comm/pair/PairResult.kt
@@ -1,0 +1,7 @@
+package info.nightscout.androidaps.plugins.pump.omnipod.dash.driver.comm.pair
+
+data class PairResult(val ltk: ByteArray) {
+    init {
+        require(ltk.size == 16)
+    }
+}
